Split base64 decoder
